+++ packages/layout-forceatlas2/src/useLayoutForceAtlas2.ts
@@ -1,4 +1,4 @@
-import forceAtlas2, { ForceAtlas2LayoutOptions } from "graphology-layout-forceatlas2";
+import forceAtlas2, { ForceAtlas2SynchronousLayoutParameters } from "graphology-layout-forceatlas2";
 
 import { useLayoutFactory } from "@react-sigma/layout-core";
 
@@ -10,4 +10,6 @@ import { useLayoutFactory } from "@react-sigma/layout-core";
  *```
  * @category Hook
  */
-export const useLayoutForceAtlas2 = useLayoutFactory<ForceAtlas2LayoutOptions>(forceAtlas2, { iterations: 100 });
+export const useLayoutForceAtlas2 = useLayoutFactory<ForceAtlas2SynchronousLayoutParameters>(forceAtlas2, {
+  iterations: 100,
+});